+++ NEWS.md
@@ -4,6 +4,7 @@
 
 # Version 1.0.0
 - 'debug' option in interpolation params to force additional output
+- Download functions DEPRECATED. User is advised to use package 'meteospain'
 
 # Version 0.9.9
 - New functions 'downloadMETEOCLIMATICcurrentday' and 'downloadMETEOCLIMATICstationlist' by MalditoBarbudo, inspired by lemuscanovas